fix: fix initial crash in build
