+++ apps/app/src/electron/game-overlay.ts
@@ -1,12 +1,10 @@
 import { BrowserWindow, screen, shell } from 'electron'
 import {
-  emitIPCGameOverlayResize,
-  onIPCGameOverlayStartIntercept,
-  onIPCGameOverlayStopIntercept,
+    emitIPCGameOverlayResize,
+    onIPCGameOverlayStartIntercept,
+    onIPCGameOverlayStopIntercept,
 } from '@/electron/events/main.ts'
 import { mouse, Point } from '@nut-tree-fork/nut-js'
-import path from 'path'
-import { EXTERNALS_DIR } from '@/electron/utils.ts'
 import { fork } from 'child_process'
 import { Deferred } from '@packages/toolbox'
 import { useGameOverlayStore } from '@/features/game-overlay/store'
@@ -15,390 +13,395 @@ import micromatch from 'micromatch'
 import { useDatabasesStore } from '@/features/databases/store'
 
 type ProcessInfo = {
-  process: string
-  pid: number
-  filepath: string
-  user: string
+    process: string
+    pid: number
+    filepath: string
+    user: string
 }
 
 type ProcessEvent = {
-  type: 'process-creation' | 'process-deletion'
-  payload: ProcessInfo
+    type: 'process-creation' | 'process-deletion'
+    payload: ProcessInfo
 }
 
 const ready = Deferred()
 
 class GameOverlay {
-  public WinControl: any = null
-  public Overlay: any = null
-  public hookedProcesses: ProcessInfo[] = []
-  public intercepting = false
-  private windows: Map<string, Electron.BrowserWindow> = new Map()
-  private markQuit = false
-  private scaleFactor = 1.0
+    public WinControl: any = null
+    public Overlay: any = null
+    public hookedProcesses: ProcessInfo[] = []
+    public intercepting = false
+    private windows: Map<string, Electron.BrowserWindow> = new Map()
+    private markQuit = false
+    private scaleFactor = 1.0
 
-  constructor() {}
+    constructor() {
+    }
 
-  public isReady = () => ready.promise
+    public isReady = () => ready.promise
 
-  public getWindow(window: string) {
-    return this.windows.get(window) || null
-  }
+    public getWindow(window: string) {
+        return this.windows.get(window) || null
+    }
 
-  public startOverlay() {
-    console.log(`[game-overlay] Starting overlay...`)
-    this.Overlay!.start()
-    // this.Overlay!.setHotkeys([
-    //   {
-    //     name: 'overlay.hotkey.toggleInputIntercept',
-    //     keyCode: 113,
-    //     modifiers: { ctrl: true },
-    //   },
-    // ])
+    public startOverlay() {
+        console.log(`[game-overlay] Starting overlay...`)
+        this.Overlay!.start()
+        // this.Overlay!.setHotkeys([
+        //   {
+        //     name: 'overlay.hotkey.toggleInputIntercept',
+        //     keyCode: 113,
+        //     modifiers: { ctrl: true },
+        //   },
+        // ])
 
-    this.Overlay!.setEventCallback((event: string, payload: any) => {
-      if (['graphics.window.event.resize', 'graphics.window'].includes(event)) {
-        const { width, height } = payload
-        emitIPCGameOverlayResize({
-          width,
-          height,
+        this.Overlay!.setEventCallback((event: string, payload: any) => {
+            if (['graphics.window.event.resize', 'graphics.window'].includes(event)) {
+                const { width, height } = payload
+                emitIPCGameOverlayResize({
+                    width,
+                    height,
+                })
+            }
+            if (event === 'game.input') {
+                const window = BrowserWindow.fromId(payload.windowId)
+                if (window) {
+                    const intpuEvent = this.Overlay!.translateInputEvent(payload)
+                    if (intpuEvent) {
+                        if ('x' in intpuEvent)
+                            intpuEvent['x'] = Math.round(intpuEvent['x'] / this.scaleFactor)
+                        if ('y' in intpuEvent)
+                            intpuEvent['y'] = Math.round(intpuEvent['y'] / this.scaleFactor)
+                        window.webContents.sendInputEvent(intpuEvent)
+                    }
+                }
+            }
+            if (event === 'game.input.intercept') {
+                this.intercepting = payload.intercepting
+            }
+            if (event === 'game.input.intercept' && payload.intercepting) {
+                const focusWin = this.windows.get('messenger-game-overlay')
+                if (focusWin) {
+                    focusWin.blurWebView()
+                    focusWin.focusOnWebView()
+                    const { top, left, right, bottom } = this.WinControl.getByPid(
+                        payload.pid,
+                    ).getDimensions()
+                    const width = right-left
+                    const height = bottom-top
+
+                    mouse.getPosition().then(async (initialPosition) => {
+                        await mouse.setPosition(
+                            new Point(left+width / 2, top+height / 2),
+                        )
+                        await mouse.leftClick()
+                        await mouse.setPosition(initialPosition)
+                    })
+                }
+            }
         })
-      }
-      if (event === 'game.input') {
-        const window = BrowserWindow.fromId(payload.windowId)
-        if (window) {
-          const intpuEvent = this.Overlay!.translateInputEvent(payload)
-          if (intpuEvent) {
-            if ('x' in intpuEvent)
-              intpuEvent['x'] = Math.round(intpuEvent['x'] / this.scaleFactor)
-            if ('y' in intpuEvent)
-              intpuEvent['y'] = Math.round(intpuEvent['y'] / this.scaleFactor)
-            window.webContents.sendInputEvent(intpuEvent)
-          }
-        }
-      }
-      if (event === 'game.input.intercept') {
-        this.intercepting = payload.intercepting
-      }
-      if (event === 'game.input.intercept' && payload.intercepting) {
-        const focusWin = this.windows.get('messenger-game-overlay')
-        if (focusWin) {
-          focusWin.blurWebView()
-          focusWin.focusOnWebView()
-          const { top, left, right, bottom } = this.WinControl.getByPid(
-            payload.pid,
-          ).getDimensions()
-          const width = right - left
-          const height = bottom - top
+    }
 
-          mouse.getPosition().then(async (initialPosition) => {
-            await mouse.setPosition(
-              new Point(left + width / 2, top + height / 2),
+    public addOverlayWindow(
+        name: string,
+        window: Electron.BrowserWindow,
+        dragBorder: number = 0,
+        captionHeight: number = 0,
+        transparent: boolean = false,
+    ) {
+        const display = screen.getDisplayNearestPoint(screen.getCursorScreenPoint())
+
+        this.Overlay!.addWindow(window.id, {
+            name,
+            transparent,
+            resizable: window.isResizable(),
+            maxWidth: window.isResizable()
+                ? display.bounds.width
+                : window.getBounds().width,
+            maxHeight: window.isResizable()
+                ? display.bounds.height
+                : window.getBounds().height,
+            minWidth: window.isResizable() ? 100 : window.getBounds().width,
+            minHeight: window.isResizable() ? 100 : window.getBounds().height,
+            nativeHandle: window.getNativeWindowHandle().readUInt32LE(0),
+            rect: {
+                x: window.getBounds().x,
+                y: window.getBounds().y,
+                width: Math.floor(window.getBounds().width * this.scaleFactor),
+                height: Math.floor(window.getBounds().height * this.scaleFactor),
+            },
+            caption: {
+                left: Math.floor(dragBorder * this.scaleFactor),
+                right: Math.floor(dragBorder * this.scaleFactor),
+                top: Math.floor(dragBorder * this.scaleFactor),
+                height: Math.floor(captionHeight * this.scaleFactor),
+            },
+            dragBorderWidth: Math.floor(dragBorder),
+        })
+
+        window.webContents.on('paint', (_, __, image: Electron.NativeImage) => {
+            if (this.markQuit) {
+                return
+            }
+            this.Overlay!.sendFrameBuffer(
+                window.id,
+                image.getBitmap(),
+                image.getSize().width,
+                image.getSize().height,
             )
-            await mouse.leftClick()
-            await mouse.setPosition(initialPosition)
-          })
+        })
+
+        window.on('ready-to-show', () => {
+            window.focusOnWebView()
+        })
+
+        window.on('resize', () => {
+            this.Overlay!.sendWindowBounds(window.id, {
+                rect: {
+                    x: window.getBounds().x,
+                    y: window.getBounds().y,
+                    width: Math.floor(window.getBounds().width * this.scaleFactor),
+                    height: Math.floor(window.getBounds().height * this.scaleFactor),
+                },
+            })
+        })
+
+        const windowId = window.id
+        window.on('closed', () => {
+            this.Overlay!.closeWindow(windowId)
+        })
+
+        window.webContents.on('cursor-changed', (_, type) => {
+            let cursor
+            switch (type) {
+                case 'default':
+                    cursor = 'IDC_ARROW'
+                    break
+                case 'pointer':
+                    cursor = 'IDC_HAND'
+                    break
+                case 'crosshair':
+                    cursor = 'IDC_CROSS'
+                    break
+                case 'text':
+                    cursor = 'IDC_IBEAM'
+                    break
+                case 'wait':
+                    cursor = 'IDC_WAIT'
+                    break
+                case 'help':
+                    cursor = 'IDC_HELP'
+                    break
+                case 'move':
+                    cursor = 'IDC_SIZEALL'
+                    break
+                case 'nwse-resize':
+                    cursor = 'IDC_SIZENWSE'
+                    break
+                case 'nesw-resize':
+                    cursor = 'IDC_SIZENESW'
+                    break
+                case 'ns-resize':
+                    cursor = 'IDC_SIZENS'
+                    break
+                case 'ew-resize':
+                    cursor = 'IDC_SIZEWE'
+                    break
+                case 'none':
+                    cursor = ''
+                    break
+            }
+            this.Overlay!.sendCommand({ command: 'cursor', cursor })
+        })
+    }
+
+    public closeAllWindows() {
+        const windows = this.windows.values()
+        for (const window of windows) {
+            window.close()
         }
-      }
-    })
-  }
-
-  public addOverlayWindow(
-    name: string,
-    window: Electron.BrowserWindow,
-    dragBorder: number = 0,
-    captionHeight: number = 0,
-    transparent: boolean = false,
-  ) {
-    const display = screen.getDisplayNearestPoint(screen.getCursorScreenPoint())
-
-    this.Overlay!.addWindow(window.id, {
-      name,
-      transparent,
-      resizable: window.isResizable(),
-      maxWidth: window.isResizable()
-        ? display.bounds.width
-        : window.getBounds().width,
-      maxHeight: window.isResizable()
-        ? display.bounds.height
-        : window.getBounds().height,
-      minWidth: window.isResizable() ? 100 : window.getBounds().width,
-      minHeight: window.isResizable() ? 100 : window.getBounds().height,
-      nativeHandle: window.getNativeWindowHandle().readUInt32LE(0),
-      rect: {
-        x: window.getBounds().x,
-        y: window.getBounds().y,
-        width: Math.floor(window.getBounds().width * this.scaleFactor),
-        height: Math.floor(window.getBounds().height * this.scaleFactor),
-      },
-      caption: {
-        left: Math.floor(dragBorder * this.scaleFactor),
-        right: Math.floor(dragBorder * this.scaleFactor),
-        top: Math.floor(dragBorder * this.scaleFactor),
-        height: Math.floor(captionHeight * this.scaleFactor),
-      },
-      dragBorderWidth: Math.floor(dragBorder),
-    })
-
-    window.webContents.on('paint', (_, __, image: Electron.NativeImage) => {
-      if (this.markQuit) {
-        return
-      }
-      this.Overlay!.sendFrameBuffer(
-        window.id,
-        image.getBitmap(),
-        image.getSize().width,
-        image.getSize().height,
-      )
-    })
-
-    window.on('ready-to-show', () => {
-      window.focusOnWebView()
-    })
-
-    window.on('resize', () => {
-      this.Overlay!.sendWindowBounds(window.id, {
-        rect: {
-          x: window.getBounds().x,
-          y: window.getBounds().y,
-          width: Math.floor(window.getBounds().width * this.scaleFactor),
-          height: Math.floor(window.getBounds().height * this.scaleFactor),
-        },
-      })
-    })
-
-    const windowId = window.id
-    window.on('closed', () => {
-      this.Overlay!.closeWindow(windowId)
-    })
-
-    window.webContents.on('cursor-changed', (_, type) => {
-      let cursor
-      switch (type) {
-        case 'default':
-          cursor = 'IDC_ARROW'
-          break
-        case 'pointer':
-          cursor = 'IDC_HAND'
-          break
-        case 'crosshair':
-          cursor = 'IDC_CROSS'
-          break
-        case 'text':
-          cursor = 'IDC_IBEAM'
-          break
-        case 'wait':
-          cursor = 'IDC_WAIT'
-          break
-        case 'help':
-          cursor = 'IDC_HELP'
-          break
-        case 'move':
-          cursor = 'IDC_SIZEALL'
-          break
-        case 'nwse-resize':
-          cursor = 'IDC_SIZENWSE'
-          break
-        case 'nesw-resize':
-          cursor = 'IDC_SIZENESW'
-          break
-        case 'ns-resize':
-          cursor = 'IDC_SIZENS'
-          break
-        case 'ew-resize':
-          cursor = 'IDC_SIZEWE'
-          break
-        case 'none':
-          cursor = ''
-          break
-      }
-      this.Overlay!.sendCommand({ command: 'cursor', cursor })
-    })
-  }
-
-  public closeAllWindows() {
-    const windows = this.windows.values()
-    for (const window of windows) {
-      window.close()
     }
-  }
 
-  public closeWindow(name: string) {
-    const window = this.windows.get(name)
-    if (window) {
-      window.close()
+    public closeWindow(name: string) {
+        const window = this.windows.get(name)
+        if (window) {
+            window.close()
+        }
     }
-  }
 
-  public injectByProcessOnceFocused(processInfo: ProcessInfo) {
-    const { pid } = processInfo
-    let timeout: ReturnType<typeof setTimeout>
-    let interval: ReturnType<typeof setInterval>
-    interval = setInterval(() => {
-      const foregroundWindow = this.WinControl.getForeground()
-      if (pid === foregroundWindow?.getPid()) {
-        clearInterval(interval)
-        clearTimeout(timeout)
-        this.injectByProcess(processInfo)
-      }
-    }, 1000)
-    setTimeout(
-      () => {
-        clearInterval(interval)
-      },
-      5 * 60 * 1000,
-    )
-  }
-
-  public injectByProcess(processInfo: ProcessInfo) {
-    for (const window of this.Overlay.getTopWindows()) {
-      if (window.processId === processInfo.pid) {
-        console.log(`[game-overlay] Injecting ${JSON.stringify(window)}`)
-        this.Overlay.injectProcess(window)
-        this.hookedProcesses.push(processInfo)
-      }
-    }
-  }
-
-  public start() {
-    const databasesStore = useDatabasesStore()
-    const gameOverlayStore = useGameOverlayStore()
-    /* Importing win-control in preload breaks reload so we import it dynamically on start instead */
-    return Promise.all([
-      import('@packages/electron-game-overlay'),
-      import('win-control'),
-      gameOverlayStore.$whenReady(),
-    ]).then(([Overlay, WinControl]) => {
-      this.WinControl = WinControl.Window
-      this.Overlay = Overlay.default
-      this.scaleFactor = screen.getDisplayNearestPoint({
-        x: 0,
-        y: 0,
-      }).scaleFactor
-
-      this.startOverlay()
-
-      watch(
-        () => [
-          gameOverlayStore.enableGameOverlay,
-          gameOverlayStore.allowlist,
-          gameOverlayStore.denylist,
-          databasesStore.data,
-        ],
-        () => {
-          if (!gameOverlayStore.enableGameOverlay) return
-          console.log('[game-overlay] Creating process watcher process')
-          const child = fork(path.join(EXTERNALS_DIR, 'detect-game.js'))
-          child.on('message', (processInfo: ProcessEvent) => {
-            if (processInfo.type === 'process-creation') {
-              const { filepath } = processInfo.payload
-              const isGame =
-                micromatch.isMatch(
-                  filepath,
-                  [
-                    ...gameOverlayStore.allowlist,
-                    ...(databasesStore.data['game-overlay-allowlist'] || []),
-                  ].filter(Boolean),
-                ) &&
-                !micromatch.isMatch(
-                  filepath,
-                  [
-                    ...gameOverlayStore.denylist,
-                    ...(databasesStore.data['game-overlay-denylist'] || []),
-                  ].filter(Boolean),
-                )
-              if (isGame) {
-                console.log('[game-overlay]: Game launched', filepath)
-                // require('windows-tlist').getProcessInfo(processInfo.payload.pid).then(({ modules }: any) => console.log(modules.map(({ path }: any) => path.substring(path.lastIndexOf('\\')+1))))
-                this.injectByProcessOnceFocused(processInfo.payload)
-              }
+    public injectByProcessOnceFocused(processInfo: ProcessInfo) {
+        const { pid } = processInfo
+        let timeout: ReturnType<typeof setTimeout>
+        let interval: ReturnType<typeof setInterval>
+        interval = setInterval(() => {
+            const foregroundWindow = this.WinControl.getForeground()
+            if (pid === foregroundWindow?.getPid()) {
+                clearInterval(interval)
+                clearTimeout(timeout)
+                this.injectByProcess(processInfo)
             }
-            if (processInfo.type === 'process-deletion') {
-              if (
-                this.hookedProcesses.find(
-                  (process) => process.pid === processInfo.payload.pid,
-                )
-              ) {
-                this.hookedProcesses = this.hookedProcesses.filter(
-                  (process) => process.pid !== processInfo.payload.pid,
-                )
-              }
+        }, 1000)
+        setTimeout(
+            () => {
+                clearInterval(interval)
+            },
+            5 * 60 * 1000,
+        )
+    }
+
+    public injectByProcess(processInfo: ProcessInfo) {
+        for (const window of this.Overlay.getTopWindows()) {
+            if (window.processId === processInfo.pid) {
+                console.log(`[game-overlay] Injecting ${ JSON.stringify(window) }`)
+                this.Overlay.injectProcess(window)
+                this.hookedProcesses.push(processInfo)
             }
-          })
-          onWatcherCleanup(() => {
-            child.kill()
-          })
-        },
-        {
-          deep: true,
-          immediate: true,
-        },
-      )
-
-      onIPCGameOverlayStartIntercept(() => {
-        this.startIntercept()
-      })
-
-      onIPCGameOverlayStopIntercept(() => {
-        this.stopIntercept()
-      })
-
-      return ready.resolve(true)
-    })
-  }
-
-  public quit() {
-    this.markQuit = true
-    this.closeAllWindows()
-
-    if (this.Overlay) {
-      this.Overlay.stop()
-    }
-  }
-
-  public startIntercept() {
-    this.Overlay!.sendCommand({
-      command: 'input.intercept',
-      intercept: true,
-    })
-  }
-
-  public stopIntercept() {
-    this.Overlay!.sendCommand({
-      command: 'input.intercept',
-      intercept: false,
-    })
-  }
-
-  public createWindow(
-    name: string,
-    option: Electron.BrowserWindowConstructorOptions,
-  ) {
-    const window = new BrowserWindow(option)
-    this.windows.set(name, window)
-
-    window.on('closed', () => {
-      this.windows.delete(name)
-    })
-
-    window.webContents.on('new-window', (e, url) => {
-      e.preventDefault()
-      shell.openExternal(url)
-    })
-
-    if (import.meta.env.DEV) {
-      window.webContents.on(
-        'before-input-event',
-        (_: Electron.Event, input: Electron.Input) => {
-          if (input.key === 'F12' && input.type === 'keyDown') {
-            window.webContents.openDevTools()
-          }
-        },
-      )
+        }
     }
 
-    return window
-  }
+    public start() {
+        const databasesStore = useDatabasesStore()
+        const gameOverlayStore = useGameOverlayStore()
+        /* Importing win-control in preload breaks reload so we import it dynamically on start instead */
+        return Promise.all([
+            require('@packages/electron-game-overlay'),
+            require('win-control'),
+            gameOverlayStore.$whenReady(),
+        ]).then(([Overlay, WinControl]) => {
+            this.WinControl = WinControl.Window
+            this.Overlay = Overlay
+            this.scaleFactor = screen.getDisplayNearestPoint({
+                x: 0,
+                y: 0,
+            }).scaleFactor
+
+            this.startOverlay()
+
+            watch(
+                () => [
+                    gameOverlayStore.enableGameOverlay,
+                    gameOverlayStore.allowlist,
+                    gameOverlayStore.denylist,
+                    databasesStore.data,
+                ],
+                () => {
+                    if (!gameOverlayStore.enableGameOverlay) return
+                    console.log('[game-overlay] Creating process watcher process')
+                    try {
+                        const child = fork(require.resolve('process-watcher'))
+                        child.on('message', (processInfo: ProcessEvent) => {
+                            if (processInfo.type === 'process-creation') {
+                                const { filepath } = processInfo.payload
+                                const isGame =
+                                    micromatch.isMatch(
+                                        filepath,
+                                        [
+                                            ...gameOverlayStore.allowlist,
+                                            ...(databasesStore.data['game-overlay-allowlist'] || []),
+                                        ].filter(Boolean),
+                                    ) &&
+                                    !micromatch.isMatch(
+                                        filepath,
+                                        [
+                                            ...gameOverlayStore.denylist,
+                                            ...(databasesStore.data['game-overlay-denylist'] || []),
+                                        ].filter(Boolean),
+                                    )
+                                if (isGame) {
+                                    console.log('[game-overlay]: Game launched', filepath)
+                                    // require('windows-tlist').getProcessInfo(processInfo.payload.pid).then(({ modules }: any) => console.log(modules.map(({ path }: any) => path.substring(path.lastIndexOf('\\')+1))))
+                                    this.injectByProcessOnceFocused(processInfo.payload)
+                                }
+                            }
+                            if (processInfo.type === 'process-deletion') {
+                                if (
+                                    this.hookedProcesses.find(
+                                        (process) => process.pid === processInfo.payload.pid,
+                                    )
+                                ) {
+                                    this.hookedProcesses = this.hookedProcesses.filter(
+                                        (process) => process.pid !== processInfo.payload.pid,
+                                    )
+                                }
+                            }
+                        })
+                        onWatcherCleanup(() => {
+                            child.kill()
+                        })
+                    } catch (e) {
+                        console.error(`[game-overlay] Couldn't create process watcher process`)
+                    }
+                },
+                {
+                    deep: true,
+                    immediate: true,
+                },
+            )
+
+            onIPCGameOverlayStartIntercept(() => {
+                this.startIntercept()
+            })
+
+            onIPCGameOverlayStopIntercept(() => {
+                this.stopIntercept()
+            })
+
+            return ready.resolve(true)
+        })
+    }
+
+    public quit() {
+        this.markQuit = true
+        this.closeAllWindows()
+
+        if (this.Overlay) {
+            this.Overlay.stop()
+        }
+    }
+
+    public startIntercept() {
+        this.Overlay!.sendCommand({
+            command: 'input.intercept',
+            intercept: true,
+        })
+    }
+
+    public stopIntercept() {
+        this.Overlay!.sendCommand({
+            command: 'input.intercept',
+            intercept: false,
+        })
+    }
+
+    public createWindow(
+        name: string,
+        option: Electron.BrowserWindowConstructorOptions,
+    ) {
+        const window = new BrowserWindow(option)
+        this.windows.set(name, window)
+
+        window.on('closed', () => {
+            this.windows.delete(name)
+        })
+
+        window.webContents.on('new-window', (e, url) => {
+            e.preventDefault()
+            shell.openExternal(url)
+        })
+
+        if (import.meta.env.DEV) {
+            window.webContents.on(
+                'before-input-event',
+                (_: Electron.Event, input: Electron.Input) => {
+                    if (input.key === 'F12' && input.type === 'keyDown') {
+                        window.webContents.openDevTools()
+                    }
+                },
+            )
+        }
+
+        return window
+    }
 }
 
 const gameOverlay = new GameOverlay()
