+++ packages/process-watcher/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "process-watcher",
+  "version": "1.0.0",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "description": "",
+  "dependencies": {
+    "wql-process-monitor": "3.0.1"
+  }
+}
